fix conflict and merge test
