--- tests/test_agent_type_exhaustiveness.py
@@ -1,11 +1,0 @@
-from pathlib import Path
-
-from bcbench.types import AgentType
-
-
-def test_all_agent_types_have_target_dir():
-    repo_path = Path("C:/test/repo")
-    for agent_type in AgentType:
-        target_dir = agent_type.get_target_dir(repo_path)
-        assert isinstance(target_dir, Path)
-        assert str(target_dir).startswith(str(repo_path))
